requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langchain-core>=0.1.10
aiohttp>=3.9.0
python-dotenv>=1.0.0
loguru>=0.7.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
requests>=2.31.0
sse-starlette>=2.0.0
